Refactor GRF Browser: Extract preview modules
Split preview functionality from main.go (2772 → 756 lines):
- preview_sprite.go: SPR/ACT preview, animation rendering
- preview_image.go: BMP/TGA/PNG/JPG preview, text and hex preview
- preview_audio.go: WAV audio playback
- preview_map.go: GAT/GND/RSW map format preview

Part of Stage 0 refactoring for ADR-012 3D Model Viewer.
