+++ cmd/grfbrowser/main.go
@@ -2,11 +2,8 @@
 package main
 
 import (
-	"bytes"
 	"flag"
 	"fmt"
-	"image"
-	"image/color"
 	_ "image/jpeg" // JPEG decoder
 	"os"
 	"path/filepath"
@@ -20,12 +17,8 @@ import (
 	"github.com/AllenDang/cimgui-go/imgui"
 	"github.com/go-gl/gl/v4.1-core/gl"
 	"github.com/gopxl/beep/v2"
-	"github.com/gopxl/beep/v2/speaker"
-	"github.com/gopxl/beep/v2/wav"
 	"github.com/sqweek/dialog"
 	_ "golang.org/x/image/bmp" // BMP decoder registration
-	"golang.org/x/text/encoding/korean"
-	"golang.org/x/text/transform"
 
 	"github.com/Faultbox/midgard-ro/pkg/formats"
 	"github.com/Faultbox/midgard-ro/pkg/grf"
@@ -750,1385 +743,6 @@ func (app *App) clearPreview() {
 
 	// Clear RSW preview (ADR-011 Stage 3)
 	app.previewRSW = nil
-}
-
-// loadSpritePreview loads a SPR file for preview.
-func (app *App) loadSpritePreview(path string) {
-	data, err := app.archive.Read(path)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading sprite: %v\n", err)
-		return
-	}
-
-	spr, err := formats.ParseSPR(data)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error parsing sprite: %v\n", err)
-		return
-	}
-
-	app.previewSPR = spr
-
-	// Create textures for all images
-	app.previewTextures = make([]*backend.Texture, len(spr.Images))
-	for i, img := range spr.Images {
-		rgba := sprImageToRGBA(&img)
-		app.previewTextures[i] = backend.NewTextureFromRgba(rgba)
-	}
-}
-
-// loadAnimationPreview loads an ACT file for preview.
-func (app *App) loadAnimationPreview(path string) {
-	data, err := app.archive.Read(path)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading animation: %v\n", err)
-		return
-	}
-
-	act, err := formats.ParseACT(data)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error parsing animation: %v\n", err)
-		return
-	}
-
-	app.previewACT = act
-
-	// Try to load corresponding SPR file (try both .spr and .SPR extensions)
-	basePath := strings.TrimSuffix(path, filepath.Ext(path))
-	sprPath := ""
-	for _, ext := range []string{".spr", ".SPR", ".Spr"} {
-		candidate := basePath + ext
-		if app.archive.Contains(candidate) {
-			sprPath = candidate
-			break
-		}
-	}
-
-	if sprPath != "" {
-		app.loadSpritePreview(sprPath)
-	} else {
-		fmt.Fprintf(os.Stderr, "SPR file not found for: %s\n", path)
-		// Debug: List files in same directory to help find the SPR
-		dir := filepath.Dir(path)
-		fmt.Fprintf(os.Stderr, "Looking for SPR files in: %s\n", dir)
-		for _, f := range app.flatFiles {
-			fNorm := strings.ReplaceAll(f, "\\", "/")
-			if strings.HasPrefix(fNorm, dir) && strings.HasSuffix(strings.ToLower(fNorm), ".spr") {
-				fmt.Fprintf(os.Stderr, "  Found SPR: %s\n", fNorm)
-			}
-		}
-	}
-
-	app.previewLastTime = time.Now()
-}
-
-// decodeTGA decodes a TGA image file.
-// Supports uncompressed true-color (type 2) and RLE compressed (type 10) TGA files,
-// which are the formats commonly used in Ragnarok Online.
-func decodeTGA(data []byte) (image.Image, error) {
-	if len(data) < 18 {
-		return nil, fmt.Errorf("TGA data too short")
-	}
-
-	// TGA header
-	idLength := int(data[0])
-	colorMapType := data[1]
-	imageType := data[2]
-	// colorMapSpec: bytes 3-7 (skip for now)
-	// imageSpec: bytes 8-17
-	xOrigin := int(data[8]) | int(data[9])<<8
-	yOrigin := int(data[10]) | int(data[11])<<8
-	width := int(data[12]) | int(data[13])<<8
-	height := int(data[14]) | int(data[15])<<8
-	bpp := int(data[16])
-	descriptor := data[17]
-
-	_ = xOrigin
-	_ = yOrigin
-
-	// Check supported formats
-	if colorMapType != 0 {
-		return nil, fmt.Errorf("color-mapped TGA not supported")
-	}
-	if imageType != 2 && imageType != 10 {
-		return nil, fmt.Errorf("unsupported TGA type %d (only uncompressed/RLE true-color supported)", imageType)
-	}
-	if bpp != 24 && bpp != 32 {
-		return nil, fmt.Errorf("unsupported TGA bit depth %d (only 24/32 supported)", bpp)
-	}
-
-	// Skip ID field
-	offset := 18 + idLength
-	if offset > len(data) {
-		return nil, fmt.Errorf("TGA data truncated")
-	}
-	pixelData := data[offset:]
-
-	// Create image
-	img := image.NewRGBA(image.Rect(0, 0, width, height))
-	bytesPerPixel := bpp / 8
-
-	// Check if image is flipped (bit 5 of descriptor = top-to-bottom)
-	topToBottom := (descriptor & 0x20) != 0
-
-	if imageType == 2 {
-		// Uncompressed
-		expectedSize := width * height * bytesPerPixel
-		if len(pixelData) < expectedSize {
-			return nil, fmt.Errorf("TGA pixel data truncated")
-		}
-
-		for y := 0; y < height; y++ {
-			destY := y
-			if !topToBottom {
-				destY = height - 1 - y
-			}
-			for x := 0; x < width; x++ {
-				i := (y*width + x) * bytesPerPixel
-				b := pixelData[i]
-				g := pixelData[i+1]
-				r := pixelData[i+2]
-				a := uint8(255)
-				if bytesPerPixel == 4 {
-					a = pixelData[i+3]
-				}
-				img.SetRGBA(x, destY, color.RGBA{R: r, G: g, B: b, A: a})
-			}
-		}
-	} else {
-		// RLE compressed (type 10)
-		pixelCount := width * height
-		pixelIdx := 0
-		dataIdx := 0
-
-		for pixelIdx < pixelCount && dataIdx < len(pixelData) {
-			packet := pixelData[dataIdx]
-			dataIdx++
-			count := int(packet&0x7F) + 1
-
-			if packet&0x80 != 0 {
-				// RLE packet - repeat single pixel
-				if dataIdx+bytesPerPixel > len(pixelData) {
-					break
-				}
-				b := pixelData[dataIdx]
-				g := pixelData[dataIdx+1]
-				r := pixelData[dataIdx+2]
-				a := uint8(255)
-				if bytesPerPixel == 4 {
-					a = pixelData[dataIdx+3]
-				}
-				dataIdx += bytesPerPixel
-
-				for i := 0; i < count && pixelIdx < pixelCount; i++ {
-					x := pixelIdx % width
-					y := pixelIdx / width
-					destY := y
-					if !topToBottom {
-						destY = height - 1 - y
-					}
-					img.SetRGBA(x, destY, color.RGBA{R: r, G: g, B: b, A: a})
-					pixelIdx++
-				}
-			} else {
-				// Raw packet - read count pixels
-				for i := 0; i < count && pixelIdx < pixelCount; i++ {
-					if dataIdx+bytesPerPixel > len(pixelData) {
-						break
-					}
-					b := pixelData[dataIdx]
-					g := pixelData[dataIdx+1]
-					r := pixelData[dataIdx+2]
-					a := uint8(255)
-					if bytesPerPixel == 4 {
-						a = pixelData[dataIdx+3]
-					}
-					dataIdx += bytesPerPixel
-
-					x := pixelIdx % width
-					y := pixelIdx / width
-					destY := y
-					if !topToBottom {
-						destY = height - 1 - y
-					}
-					img.SetRGBA(x, destY, color.RGBA{R: r, G: g, B: b, A: a})
-					pixelIdx++
-				}
-			}
-		}
-	}
-
-	return img, nil
-}
-
-// loadImagePreview loads an image file (BMP, TGA, JPG, PNG) for preview.
-func (app *App) loadImagePreview(path string) {
-	data, err := app.archive.Read(path)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading image: %v\n", err)
-		return
-	}
-
-	// Decode image - use image.Decode which auto-detects format
-	// BMP is registered via golang.org/x/image/bmp import
-	// JPEG and PNG are in standard library
-	var img image.Image
-	ext := strings.ToLower(filepath.Ext(path))
-
-	if ext == ".tga" {
-		// TGA needs special handling (not in standard library)
-		img, err = decodeTGA(data)
-	} else {
-		img, _, err = image.Decode(bytes.NewReader(data))
-	}
-
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error decoding image %s: %v\n", ext, err)
-		return
-	}
-
-	// Convert to RGBA
-	bounds := img.Bounds()
-	rgba := image.NewRGBA(bounds)
-	for y := bounds.Min.Y; y < bounds.Max.Y; y++ {
-		for x := bounds.Min.X; x < bounds.Max.X; x++ {
-			rgba.Set(x, y, img.At(x, y))
-		}
-	}
-
-	// Create texture
-	app.previewImage = backend.NewTextureFromRgba(rgba)
-	app.previewImgSize = [2]int{bounds.Dx(), bounds.Dy()}
-}
-
-// loadTextPreview loads a text file for preview.
-func (app *App) loadTextPreview(path string) {
-	data, err := app.archive.Read(path)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading text file: %v\n", err)
-		return
-	}
-
-	// Try to convert from EUC-KR to UTF-8 if it looks like Korean
-	text := string(data)
-	if hasHighBytes(data) {
-		decoder := korean.EUCKR.NewDecoder()
-		if decoded, _, err := transform.String(decoder, text); err == nil {
-			text = decoded
-		}
-	}
-
-	// Limit preview size to avoid performance issues
-	const maxPreviewSize = 64 * 1024 // 64KB
-	if len(text) > maxPreviewSize {
-		text = text[:maxPreviewSize] + "\n\n... (truncated)"
-	}
-
-	app.previewText = text
-}
-
-// loadHexPreview loads raw bytes for hex preview.
-func (app *App) loadHexPreview(path string) {
-	data, err := app.archive.Read(path)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading file: %v\n", err)
-		return
-	}
-
-	app.previewHexSize = int64(len(data))
-
-	// Limit hex preview to first 4KB
-	const maxHexSize = 4 * 1024
-	if len(data) > maxHexSize {
-		app.previewHex = data[:maxHexSize]
-	} else {
-		app.previewHex = data
-	}
-}
-
-// hasHighBytes checks if data contains non-ASCII bytes (potential EUC-KR).
-func hasHighBytes(data []byte) bool {
-	for _, b := range data {
-		if b > 127 {
-			return true
-		}
-	}
-	return false
-}
-
-// renderSpritePreview renders the sprite preview with frame navigation.
-func (app *App) renderSpritePreview() {
-	if app.previewSPR == nil || len(app.previewTextures) == 0 {
-		imgui.TextDisabled("Failed to load sprite")
-		return
-	}
-
-	spr := app.previewSPR
-	imgui.Text(fmt.Sprintf("Version: %s", spr.Version))
-	imgui.Text(fmt.Sprintf("Images: %d", len(spr.Images)))
-
-	if spr.Palette != nil {
-		imgui.Text("Palette: Yes (256 colors)")
-	}
-
-	imgui.Separator()
-
-	// Frame navigation
-	imgui.Text(fmt.Sprintf("Frame: %d / %d", app.previewFrame+1, len(spr.Images)))
-	imgui.SameLine()
-
-	if imgui.Button("<") && app.previewFrame > 0 {
-		app.previewFrame--
-	}
-	imgui.SameLine()
-	if imgui.Button(">") && app.previewFrame < len(spr.Images)-1 {
-		app.previewFrame++
-	}
-
-	// Zoom controls
-	imgui.SameLine()
-	imgui.Text("  Zoom:")
-	imgui.SameLine()
-	if imgui.Button("-") && app.previewZoom > 0.5 {
-		app.previewZoom -= 0.5
-	}
-	imgui.SameLine()
-	imgui.Text(fmt.Sprintf("%.1fx", app.previewZoom))
-	imgui.SameLine()
-	if imgui.Button("+") && app.previewZoom < 8.0 {
-		app.previewZoom += 0.5
-	}
-
-	imgui.Separator()
-
-	// Display current frame centered in available space
-	if app.previewFrame < len(app.previewTextures) {
-		tex := app.previewTextures[app.previewFrame]
-		if tex != nil {
-			img := spr.Images[app.previewFrame]
-			w := float32(img.Width) * app.previewZoom
-			h := float32(img.Height) * app.previewZoom
-
-			// Center the image both horizontally and vertically
-			avail := imgui.ContentRegionAvail()
-			startX := imgui.CursorPosX()
-			startY := imgui.CursorPosY()
-			if w < avail.X {
-				imgui.SetCursorPosX(startX + (avail.X-w)/2)
-			}
-			if h < avail.Y {
-				imgui.SetCursorPosY(startY + (avail.Y-h)/2)
-			}
-
-			// Draw with checkerboard background to show transparency
-			imgui.ImageWithBgV(
-				tex.ID,
-				imgui.NewVec2(w, h),
-				imgui.NewVec2(0, 0),
-				imgui.NewVec2(1, 1),
-				imgui.NewVec4(0.2, 0.2, 0.2, 1.0), // Dark gray background
-				imgui.NewVec4(1, 1, 1, 1),         // White tint (no tint)
-			)
-		}
-	}
-}
-
-// renderAnimationPreview renders the animation preview (frame display only, controls in Actions panel).
-func (app *App) renderAnimationPreview() {
-	if app.previewACT == nil {
-		imgui.TextDisabled("Failed to load animation")
-		return
-	}
-
-	act := app.previewACT
-
-	// Update animation timing
-	if len(act.Actions) > 0 {
-		action := act.Actions[app.previewAction]
-
-		if app.previewPlaying && len(action.Frames) > 0 {
-			// Get interval from ACT file (default 4 ticks if not specified)
-			// ACT intervals are in game ticks; multiply by 24ms per tick for real time
-			interval := float32(4.0) // default 4 ticks
-			if app.previewAction < len(act.Intervals) && act.Intervals[app.previewAction] > 0 {
-				interval = act.Intervals[app.previewAction]
-			}
-
-			// Convert ticks to milliseconds (24ms per tick is standard RO timing)
-			// Apply minimum floor of 100ms for readability
-			intervalMs := interval * 24.0
-			if intervalMs < 100.0 {
-				intervalMs = 100.0
-			}
-
-			elapsed := time.Since(app.previewLastTime).Milliseconds()
-			if elapsed >= int64(intervalMs) {
-				app.previewFrame = (app.previewFrame + 1) % len(action.Frames)
-				app.previewLastTime = time.Now()
-			}
-		}
-
-		// Render current frame layers
-		if app.previewFrame < len(action.Frames) && app.previewSPR != nil {
-			frame := action.Frames[app.previewFrame]
-			app.renderACTFrame(&frame)
-		} else if app.previewSPR == nil {
-			imgui.TextDisabled("No sprite loaded (SPR file not found)")
-		}
-	}
-}
-
-// renderActionsPanel renders the Actions panel for ACT files.
-func (app *App) renderActionsPanel() {
-	if app.previewACT == nil {
-		return
-	}
-
-	act := app.previewACT
-
-	// Playback controls at top
-	if app.previewPlaying {
-		if imgui.ButtonV("Pause", imgui.NewVec2(-1, 0)) {
-			app.previewPlaying = false
-		}
-	} else {
-		if imgui.ButtonV("Play", imgui.NewVec2(-1, 0)) {
-			app.previewPlaying = true
-			app.previewLastTime = time.Now()
-		}
-	}
-
-	imgui.Text("(Space to toggle)")
-
-	imgui.Separator()
-
-	// Frame info
-	if len(act.Actions) > 0 && app.previewAction < len(act.Actions) {
-		action := act.Actions[app.previewAction]
-		imgui.Text(fmt.Sprintf("Frame: %d / %d", app.previewFrame+1, len(action.Frames)))
-
-		// Frame navigation
-		if imgui.Button("<##frame") && app.previewFrame > 0 {
-			app.previewFrame--
-		}
-		imgui.SameLine()
-		if imgui.Button(">##frame") && app.previewFrame < len(action.Frames)-1 {
-			app.previewFrame++
-		}
-	}
-
-	imgui.Separator()
-	imgui.Text("Actions:")
-
-	// Scrollable action list
-	totalActions := len(act.Actions)
-	if imgui.BeginChildStrV("ActionList", imgui.NewVec2(0, 0), imgui.ChildFlagsBorders, 0) {
-		for i := 0; i < totalActions; i++ {
-			action := act.Actions[i]
-			name := getActionName(i, totalActions)
-			label := fmt.Sprintf("%d: %s (%d)", i, name, len(action.Frames))
-
-			isSelected := i == app.previewAction
-			if imgui.SelectableBoolV(label, isSelected, 0, imgui.NewVec2(0, 0)) {
-				app.previewAction = i
-				app.previewFrame = 0
-			}
-		}
-	}
-	imgui.EndChild()
-}
-
-// Direction names for 8-direction sprites (standard RO order).
-var directionNames = []string{
-	"S", "SW", "W", "NW", "N", "NE", "E", "SE",
-}
-
-// Monster action type names (typically 5-8 action types with 8 directions each).
-// Monsters have: Idle, Walk, Attack, Damage, Die (and sometimes more attacks).
-var monsterActionNames = []string{
-	"Idle",     // 0
-	"Walk",     // 1
-	"Attack",   // 2
-	"Damage",   // 3
-	"Die",      // 4
-	"Attack 2", // 5 (some monsters)
-	"Attack 3", // 6 (some monsters)
-	"Special",  // 7 (some monsters)
-}
-
-// Player action type names (typically 13+ action types with 8 directions each).
-// Players have more actions including Sit, Pick Up, Standby, multiple attacks, casting.
-var playerActionNames = []string{
-	"Idle",        // 0
-	"Walk",        // 1
-	"Sit",         // 2
-	"Pick Up",     // 3
-	"Standby",     // 4
-	"Attack 1",    // 5
-	"Damage",      // 6
-	"Die",         // 7
-	"Dead",        // 8
-	"Attack 2",    // 9
-	"Attack 3",    // 10
-	"Skill Cast",  // 11
-	"Skill Ready", // 12
-	"Freeze",      // 13
-}
-
-// getActionName returns a standard RO action name for the given index.
-// For sprites with 8 directions per action, shows "ActionType Dir" format.
-// Uses different naming based on total action count to detect monster vs player sprites.
-func getActionName(index, totalActions int) string {
-	// Check if this looks like an 8-direction sprite (multiple of 8 actions)
-	if totalActions >= 8 && totalActions%8 == 0 {
-		actionType := index / 8
-		direction := index % 8
-		dirName := directionNames[direction]
-
-		// Determine sprite type based on action count:
-		// - Monster sprites: typically 40-64 actions (5-8 action types)
-		// - Player sprites: typically 80+ actions (10+ action types)
-		var typeName string
-		actionTypeCount := totalActions / 8
-
-		if actionTypeCount <= 8 {
-			// Likely a monster sprite
-			if actionType < len(monsterActionNames) {
-				typeName = monsterActionNames[actionType]
-			} else {
-				typeName = fmt.Sprintf("Action%d", actionType)
-			}
-		} else {
-			// Likely a player sprite
-			if actionType < len(playerActionNames) {
-				typeName = playerActionNames[actionType]
-			} else {
-				typeName = fmt.Sprintf("Action%d", actionType)
-			}
-		}
-
-		return fmt.Sprintf("%s %s", typeName, dirName)
-	}
-
-	// Non-8-direction sprite (items, effects, etc.) - just show index
-	return fmt.Sprintf("Action %d", index)
-}
-
-// renderACTFrame renders a single ACT frame with all its layers.
-func (app *App) renderACTFrame(frame *formats.Frame) {
-	if len(frame.Layers) == 0 {
-		imgui.TextDisabled("Empty frame")
-		return
-	}
-
-	// For now, just render the first valid layer's sprite
-	validLayerFound := false
-	allLayersEmpty := true
-
-	for _, layer := range frame.Layers {
-		if layer.SpriteID >= 0 {
-			allLayersEmpty = false
-			break
-		}
-	}
-
-	// Show informative message for empty frames (common in garment/accessory ACT files)
-	if allLayersEmpty {
-		imgui.TextDisabled("Frame has no sprites")
-		imgui.TextDisabled("(Accessory/garment overlay - uses base sprite)")
-		return
-	}
-
-	for _, layer := range frame.Layers {
-		if layer.SpriteID < 0 {
-			continue
-		}
-
-		// Calculate actual sprite index based on sprite type
-		// Type 0 = indexed (palette), Type 1 = RGBA (true-color)
-		// RGBA sprites are stored after indexed sprites in the SPR file
-		spriteIndex := int(layer.SpriteID)
-		if layer.SpriteType == 1 && app.previewSPR != nil {
-			spriteIndex += app.previewSPR.IndexedCount
-		}
-
-		if spriteIndex >= len(app.previewTextures) {
-			continue
-		}
-
-		tex := app.previewTextures[spriteIndex]
-		if tex == nil {
-			continue
-		}
-
-		validLayerFound = true
-		img := app.previewSPR.Images[spriteIndex]
-		w := float32(img.Width) * app.previewZoom * layer.ScaleX
-		h := float32(img.Height) * app.previewZoom * layer.ScaleY
-
-		// Center the image both horizontally and vertically
-		avail := imgui.ContentRegionAvail()
-		startX := imgui.CursorPosX()
-		startY := imgui.CursorPosY()
-		if w < avail.X {
-			imgui.SetCursorPosX(startX + (avail.X-w)/2)
-		}
-		if h < avail.Y {
-			imgui.SetCursorPosY(startY + (avail.Y-h)/2)
-		}
-
-		// Apply layer color tint
-		tint := imgui.NewVec4(
-			float32(layer.Color[0])/255.0,
-			float32(layer.Color[1])/255.0,
-			float32(layer.Color[2])/255.0,
-			float32(layer.Color[3])/255.0,
-		)
-
-		imgui.ImageWithBgV(
-			tex.ID,
-			imgui.NewVec2(w, h),
-			imgui.NewVec2(0, 0),
-			imgui.NewVec2(1, 1),
-			imgui.NewVec4(0.2, 0.2, 0.2, 1.0),
-			tint,
-		)
-
-		// Only render first valid layer for now (proper compositing would need DrawList)
-		break
-	}
-
-	if !validLayerFound {
-		imgui.TextDisabled("No renderable sprites in frame")
-	}
-}
-
-// renderImagePreview renders an image (BMP, TGA, JPG, PNG) with zoom controls.
-func (app *App) renderImagePreview() {
-	if app.previewImage == nil {
-		imgui.TextDisabled("Failed to load image")
-		return
-	}
-
-	imgui.Text(fmt.Sprintf("Size: %d x %d", app.previewImgSize[0], app.previewImgSize[1]))
-
-	// Zoom controls
-	imgui.Text("Zoom:")
-	imgui.SameLine()
-	if imgui.Button("-##imgzoom") && app.previewZoom > 0.25 {
-		app.previewZoom -= 0.25
-	}
-	imgui.SameLine()
-	imgui.Text(fmt.Sprintf("%.0f%%", app.previewZoom*100))
-	imgui.SameLine()
-	if imgui.Button("+##imgzoom") && app.previewZoom < 4.0 {
-		app.previewZoom += 0.25
-	}
-	imgui.SameLine()
-	if imgui.Button("Reset##imgzoom") {
-		app.previewZoom = 1.0
-	}
-
-	imgui.Separator()
-
-	// Display image centered
-	w := float32(app.previewImgSize[0]) * app.previewZoom
-	h := float32(app.previewImgSize[1]) * app.previewZoom
-
-	avail := imgui.ContentRegionAvail()
-	startX := imgui.CursorPosX()
-	startY := imgui.CursorPosY()
-	if w < avail.X {
-		imgui.SetCursorPosX(startX + (avail.X-w)/2)
-	}
-	if h < avail.Y {
-		imgui.SetCursorPosY(startY + (avail.Y-h)/2)
-	}
-
-	imgui.ImageWithBgV(
-		app.previewImage.ID,
-		imgui.NewVec2(w, h),
-		imgui.NewVec2(0, 0),
-		imgui.NewVec2(1, 1),
-		imgui.NewVec4(0.2, 0.2, 0.2, 1.0),
-		imgui.NewVec4(1, 1, 1, 1),
-	)
-}
-
-// renderTextPreview renders a text file with scrolling.
-func (app *App) renderTextPreview() {
-	if app.previewText == "" {
-		imgui.TextDisabled("Empty file or failed to load")
-		return
-	}
-
-	imgui.Text(fmt.Sprintf("Size: %d bytes", len(app.previewText)))
-	imgui.Separator()
-
-	// Scrollable text area
-	flags := imgui.WindowFlagsHorizontalScrollbar
-	if imgui.BeginChildStrV("TextPreview", imgui.NewVec2(0, 0), imgui.ChildFlagsBorders, flags) {
-		imgui.TextUnformatted(app.previewText)
-	}
-	imgui.EndChild()
-}
-
-// renderHexPreview renders a hex dump of binary data.
-func (app *App) renderHexPreview() {
-	if app.previewHex == nil {
-		imgui.TextDisabled("Failed to load file")
-		return
-	}
-
-	imgui.Text(fmt.Sprintf("File size: %d bytes", app.previewHexSize))
-	if int64(len(app.previewHex)) < app.previewHexSize {
-		imgui.SameLine()
-		imgui.TextDisabled(fmt.Sprintf("(showing first %d bytes)", len(app.previewHex)))
-	}
-	imgui.Separator()
-
-	// Scrollable hex view
-	flags := imgui.WindowFlagsHorizontalScrollbar
-	if imgui.BeginChildStrV("HexPreview", imgui.NewVec2(0, 0), imgui.ChildFlagsBorders, flags) {
-		// Render hex dump in classic format: offset | hex bytes | ascii
-		const bytesPerLine = 16
-		for offset := 0; offset < len(app.previewHex); offset += bytesPerLine {
-			// Offset
-			line := fmt.Sprintf("%08X  ", offset)
-
-			// Hex bytes
-			for i := 0; i < bytesPerLine; i++ {
-				if offset+i < len(app.previewHex) {
-					line += fmt.Sprintf("%02X ", app.previewHex[offset+i])
-				} else {
-					line += "   "
-				}
-				if i == 7 {
-					line += " "
-				}
-			}
-
-			// ASCII representation
-			line += " |"
-			for i := 0; i < bytesPerLine && offset+i < len(app.previewHex); i++ {
-				b := app.previewHex[offset+i]
-				if b >= 32 && b < 127 {
-					line += string(b)
-				} else {
-					line += "."
-				}
-			}
-			line += "|"
-
-			imgui.Text(line)
-		}
-	}
-	imgui.EndChild()
-}
-
-// loadAudioPreview loads a WAV file for audio preview.
-func (app *App) loadAudioPreview(path string) {
-	data, err := app.archive.Read(path)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading audio file: %v\n", err)
-		return
-	}
-
-	// Decode WAV from memory
-	streamer, format, err := wav.Decode(bytes.NewReader(data))
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error decoding WAV: %v\n", err)
-		return
-	}
-
-	// Initialize speaker once (use common sample rate)
-	speakerInitOnce.Do(func() {
-		err := speaker.Init(format.SampleRate, format.SampleRate.N(time.Second/10))
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error initializing speaker: %v\n", err)
-			return
-		}
-		speakerInited = true
-	})
-
-	if !speakerInited {
-		streamer.Close()
-		return
-	}
-
-	app.audioStreamer = streamer
-	app.audioFormat = format
-	app.audioLength = streamer.Len()
-	app.audioSampleRate = format.SampleRate
-	app.audioPlaying = false
-	app.audioCtrl = nil
-}
-
-// renderAudioPreview renders the audio player with controls.
-func (app *App) renderAudioPreview() {
-	if app.audioStreamer == nil {
-		imgui.TextDisabled("Failed to load audio")
-		return
-	}
-
-	// Audio info
-	duration := app.audioSampleRate.D(app.audioLength)
-	imgui.Text(fmt.Sprintf("Format: %d Hz, %d ch", app.audioFormat.SampleRate, app.audioFormat.NumChannels))
-	imgui.Text(fmt.Sprintf("Duration: %.1f sec", duration.Seconds()))
-
-	imgui.Separator()
-
-	// Play/Stop buttons
-	if app.audioPlaying {
-		if imgui.ButtonV("Stop", imgui.NewVec2(80, 0)) {
-			app.stopAudio()
-		}
-	} else {
-		if imgui.ButtonV("Play", imgui.NewVec2(80, 0)) {
-			app.playAudio()
-		}
-	}
-
-	imgui.SameLine()
-
-	// Progress bar
-	var progress float32
-	var currentPos int
-	if app.audioStreamer != nil && app.audioLength > 0 {
-		currentPos = app.audioStreamer.Position()
-		progress = float32(currentPos) / float32(app.audioLength)
-	}
-
-	currentTime := app.audioSampleRate.D(currentPos)
-	imgui.Text(fmt.Sprintf("%.1f / %.1f", currentTime.Seconds(), duration.Seconds()))
-
-	// Progress bar (full width)
-	imgui.ProgressBarV(progress, imgui.NewVec2(-1, 0), "")
-
-	// Check if playback finished
-	if app.audioPlaying && currentPos >= app.audioLength {
-		app.audioPlaying = false
-	}
-}
-
-// playAudio starts audio playback.
-func (app *App) playAudio() {
-	if app.audioStreamer == nil || !speakerInited {
-		return
-	}
-
-	// Reset to beginning
-	if err := app.audioStreamer.Seek(0); err != nil {
-		fmt.Fprintf(os.Stderr, "Error seeking audio: %v\n", err)
-		return
-	}
-
-	// Create control wrapper for pause/resume
-	app.audioCtrl = &beep.Ctrl{Streamer: app.audioStreamer, Paused: false}
-	app.audioPlaying = true
-
-	// Play with callback when done
-	speaker.Play(beep.Seq(app.audioCtrl, beep.Callback(func() {
-		app.audioPlaying = false
-	})))
-}
-
-// stopAudio stops audio playback and releases resources.
-func (app *App) stopAudio() {
-	if speakerInited {
-		speaker.Clear()
-	}
-	app.audioPlaying = false
-	app.audioCtrl = nil
-	if app.audioStreamer != nil {
-		app.audioStreamer.Close()
-		app.audioStreamer = nil
-	}
-}
-
-// loadGATPreview loads a GAT file for preview.
-func (app *App) loadGATPreview(path string) {
-	data, err := app.archive.Read(path)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading GAT file: %v\n", err)
-		return
-	}
-
-	gat, err := formats.ParseGAT(data)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error parsing GAT: %v\n", err)
-		return
-	}
-
-	app.previewGAT = gat
-	app.previewGATZoom = 1.0
-
-	// Create visualization texture
-	app.createGATTexture()
-}
-
-// createGATTexture creates a texture from GAT data for visualization.
-func (app *App) createGATTexture() {
-	if app.previewGAT == nil {
-		return
-	}
-
-	gat := app.previewGAT
-	width := int(gat.Width)
-	height := int(gat.Height)
-
-	// Create RGBA image
-	rgba := image.NewRGBA(image.Rect(0, 0, width, height))
-
-	// Color map for cell types
-	for y := 0; y < height; y++ {
-		for x := 0; x < width; x++ {
-			cell := gat.GetCell(x, y)
-			if cell == nil {
-				continue
-			}
-
-			var c color.RGBA
-			switch cell.Type {
-			case formats.GATWalkable:
-				c = color.RGBA{R: 100, G: 200, B: 100, A: 255} // Green
-			case formats.GATBlocked:
-				c = color.RGBA{R: 200, G: 80, B: 80, A: 255} // Red
-			case formats.GATWater:
-				c = color.RGBA{R: 80, G: 150, B: 220, A: 255} // Blue
-			case formats.GATWalkableWater:
-				c = color.RGBA{R: 100, G: 200, B: 200, A: 255} // Cyan
-			case formats.GATSnipeable:
-				c = color.RGBA{R: 220, G: 200, B: 80, A: 255} // Yellow
-			case formats.GATBlockedSnipe:
-				c = color.RGBA{R: 220, G: 150, B: 80, A: 255} // Orange
-			default:
-				c = color.RGBA{R: 128, G: 128, B: 128, A: 255} // Gray
-			}
-
-			// Flip Y for display (GAT origin is bottom-left)
-			rgba.SetRGBA(x, height-1-y, c)
-		}
-	}
-
-	// Create texture
-	app.previewGATTex = backend.NewTextureFromRgba(rgba)
-}
-
-// renderGATPreview renders the GAT visualization.
-func (app *App) renderGATPreview() {
-	if app.previewGAT == nil {
-		imgui.TextDisabled("Failed to load GAT file")
-		return
-	}
-
-	gat := app.previewGAT
-
-	// Info
-	imgui.Text(fmt.Sprintf("Size: %d x %d cells", gat.Width, gat.Height))
-	imgui.Text(fmt.Sprintf("Version: %s", gat.Version))
-
-	// Cell type counts
-	counts := gat.CountByType()
-	imgui.Text(fmt.Sprintf("Walkable: %d | Blocked: %d | Water: %d",
-		counts[formats.GATWalkable]+counts[formats.GATWalkableWater],
-		counts[formats.GATBlocked]+counts[formats.GATBlockedSnipe],
-		counts[formats.GATWater]+counts[formats.GATWalkableWater]))
-
-	// Altitude range
-	minAlt, maxAlt := gat.GetAltitudeRange()
-	imgui.Text(fmt.Sprintf("Altitude: %.1f to %.1f", minAlt, maxAlt))
-
-	imgui.Separator()
-
-	// Zoom controls (buttons only, keyboard +/- handled globally in handleKeyboardShortcuts)
-	imgui.Text("Zoom:")
-	imgui.SameLine()
-	if imgui.Button("-##gatzoom") && app.previewGATZoom > 0.25 {
-		app.previewGATZoom -= 0.25
-	}
-	imgui.SameLine()
-	imgui.Text(fmt.Sprintf("%.0f%%", app.previewGATZoom*100))
-	imgui.SameLine()
-	if imgui.Button("+##gatzoom") && app.previewGATZoom < 8.0 {
-		app.previewGATZoom += 0.25
-	}
-	imgui.SameLine()
-	if imgui.Button("Fit##gatzoom") {
-		// Calculate zoom to fit in available space
-		avail := imgui.ContentRegionAvail()
-		zoomX := avail.X / float32(gat.Width)
-		zoomY := avail.Y / float32(gat.Height)
-		app.previewGATZoom = min(zoomX, zoomY)
-		if app.previewGATZoom < 0.1 {
-			app.previewGATZoom = 0.1
-		}
-	}
-
-	imgui.Separator()
-
-	// Legend
-	imgui.TextColored(imgui.NewVec4(0.4, 0.8, 0.4, 1), "Walkable")
-	imgui.SameLine()
-	imgui.TextColored(imgui.NewVec4(0.8, 0.3, 0.3, 1), "Blocked")
-	imgui.SameLine()
-	imgui.TextColored(imgui.NewVec4(0.3, 0.6, 0.9, 1), "Water")
-	imgui.SameLine()
-	imgui.TextColored(imgui.NewVec4(0.9, 0.8, 0.3, 1), "Snipeable")
-
-	imgui.Separator()
-
-	// Display texture
-	if app.previewGATTex != nil {
-		w := float32(gat.Width) * app.previewGATZoom
-		h := float32(gat.Height) * app.previewGATZoom
-
-		// Scrollable child region for large maps
-		if imgui.BeginChildStrV("GATView", imgui.NewVec2(0, 0), imgui.ChildFlagsBorders, imgui.WindowFlagsHorizontalScrollbar) {
-			imgui.ImageWithBgV(
-				app.previewGATTex.ID,
-				imgui.NewVec2(w, h),
-				imgui.NewVec2(0, 0),
-				imgui.NewVec2(1, 1),
-				imgui.NewVec4(0.1, 0.1, 0.1, 1.0), // Dark background
-				imgui.NewVec4(1, 1, 1, 1),         // No tint
-			)
-		}
-		imgui.EndChild()
-	}
-}
-
-// loadGNDPreview loads a GND file for preview.
-func (app *App) loadGNDPreview(path string) {
-	data, err := app.archive.Read(path)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading GND file: %v\n", err)
-		return
-	}
-
-	gnd, err := formats.ParseGND(data)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error parsing GND: %v\n", err)
-		return
-	}
-
-	app.previewGND = gnd
-	app.previewGNDZoom = 1.0
-
-	// Create height map visualization texture
-	app.createGNDTexture()
-}
-
-// createGNDTexture creates a height map texture from GND data for visualization.
-func (app *App) createGNDTexture() {
-	if app.previewGND == nil {
-		return
-	}
-
-	gnd := app.previewGND
-	width := int(gnd.Width)
-	height := int(gnd.Height)
-
-	// Create RGBA image - height map visualization
-	rgba := image.NewRGBA(image.Rect(0, 0, width, height))
-
-	// Get altitude range for normalization
-	minAlt, maxAlt := gnd.GetAltitudeRange()
-	altRange := maxAlt - minAlt
-	if altRange == 0 {
-		altRange = 1 // Avoid division by zero
-	}
-
-	// Color map: low altitude = dark blue, high altitude = white/yellow
-	for y := 0; y < height; y++ {
-		for x := 0; x < width; x++ {
-			tile := gnd.GetTile(x, y)
-			if tile == nil {
-				continue
-			}
-
-			// Average altitude of the tile
-			avgAlt := (tile.Altitude[0] + tile.Altitude[1] + tile.Altitude[2] + tile.Altitude[3]) / 4.0
-
-			// Normalize to 0-1 (note: RO uses negative Y for height)
-			normalized := (maxAlt - avgAlt) / altRange
-
-			// Color gradient: blue (low) -> green -> yellow -> white (high)
-			var r, g, b uint8
-			if normalized < 0.25 {
-				// Dark blue to blue
-				t := normalized * 4
-				r = uint8(20 * t)
-				g = uint8(50 + 50*t)
-				b = uint8(100 + 100*t)
-			} else if normalized < 0.5 {
-				// Blue to green
-				t := (normalized - 0.25) * 4
-				r = uint8(20 + 80*t)
-				g = uint8(100 + 100*t)
-				b = uint8(200 - 100*t)
-			} else if normalized < 0.75 {
-				// Green to yellow
-				t := (normalized - 0.5) * 4
-				r = uint8(100 + 155*t)
-				g = uint8(200 + 55*t)
-				b = uint8(100 - 50*t)
-			} else {
-				// Yellow to white
-				t := (normalized - 0.75) * 4
-				r = uint8(255)
-				g = uint8(255)
-				b = uint8(50 + 205*t)
-			}
-
-			// Highlight tiles with no top surface (cliffs/holes)
-			if tile.TopSurface < 0 {
-				r, g, b = 80, 20, 20 // Dark red for no surface
-			}
-
-			rgba.Set(x, y, color.RGBA{R: r, G: g, B: b, A: 255})
-		}
-	}
-
-	// Create texture
-	app.previewGNDTex = backend.NewTextureFromRgba(rgba)
-}
-
-// renderGNDPreview renders the GND visualization.
-func (app *App) renderGNDPreview() {
-	if app.previewGND == nil {
-		imgui.TextDisabled("Failed to load GND file")
-		return
-	}
-
-	gnd := app.previewGND
-
-	// Info
-	imgui.Text(fmt.Sprintf("Size: %d x %d tiles", gnd.Width, gnd.Height))
-	imgui.Text(fmt.Sprintf("Version: %s", gnd.Version))
-	imgui.Text(fmt.Sprintf("Zoom Factor: %.1f", gnd.Zoom))
-	imgui.Text(fmt.Sprintf("Textures: %d", len(gnd.Textures)))
-	imgui.Text(fmt.Sprintf("Surfaces: %d", len(gnd.Surfaces)))
-	imgui.Text(fmt.Sprintf("Lightmaps: %d (%dx%d)", len(gnd.Lightmaps), gnd.LightmapWidth, gnd.LightmapHeight))
-
-	// Altitude range
-	minAlt, maxAlt := gnd.GetAltitudeRange()
-	imgui.Text(fmt.Sprintf("Altitude: %.1f to %.1f", minAlt, maxAlt))
-
-	imgui.Separator()
-
-	// Zoom controls
-	imgui.Text("Zoom:")
-	imgui.SameLine()
-	if imgui.Button("-##gndzoom") && app.previewGNDZoom > 0.25 {
-		app.previewGNDZoom -= 0.25
-	}
-	imgui.SameLine()
-	imgui.Text(fmt.Sprintf("%.0f%%", app.previewGNDZoom*100))
-	imgui.SameLine()
-	if imgui.Button("+##gndzoom") && app.previewGNDZoom < 8.0 {
-		app.previewGNDZoom += 0.25
-	}
-	imgui.SameLine()
-	if imgui.Button("Fit##gndzoom") {
-		avail := imgui.ContentRegionAvail()
-		zoomX := avail.X / float32(gnd.Width)
-		zoomY := avail.Y / float32(gnd.Height)
-		app.previewGNDZoom = min(zoomX, zoomY)
-		if app.previewGNDZoom < 0.1 {
-			app.previewGNDZoom = 0.1
-		}
-	}
-
-	imgui.Separator()
-
-	// Texture list (collapsible)
-	if imgui.TreeNodeExStrV("Textures", imgui.TreeNodeFlagsNone) {
-		for i, tex := range gnd.Textures {
-			imgui.Text(fmt.Sprintf("%d: %s", i, tex))
-		}
-		imgui.TreePop()
-	}
-
-	imgui.Separator()
-
-	// Display height map texture
-	if app.previewGNDTex != nil {
-		w := float32(gnd.Width) * app.previewGNDZoom
-		h := float32(gnd.Height) * app.previewGNDZoom
-
-		// Scrollable child region for large maps
-		if imgui.BeginChildStrV("GNDView", imgui.NewVec2(0, 0), imgui.ChildFlagsBorders, imgui.WindowFlagsHorizontalScrollbar) {
-			imgui.ImageWithBgV(
-				app.previewGNDTex.ID,
-				imgui.NewVec2(w, h),
-				imgui.NewVec2(0, 0),
-				imgui.NewVec2(1, 1),
-				imgui.NewVec4(0.1, 0.1, 0.1, 1.0), // Dark background
-				imgui.NewVec4(1, 1, 1, 1),         // No tint
-			)
-		}
-		imgui.EndChild()
-	}
-}
-
-// loadRSWPreview loads a RSW file for preview.
-func (app *App) loadRSWPreview(path string) {
-	data, err := app.archive.Read(path)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading RSW file: %v\n", err)
-		return
-	}
-
-	rsw, err := formats.ParseRSW(data)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error parsing RSW: %v\n", err)
-		return
-	}
-
-	app.previewRSW = rsw
-}
-
-// renderRSWPreview renders the RSW world info panel.
-func (app *App) renderRSWPreview() {
-	if app.previewRSW == nil {
-		imgui.TextDisabled("Failed to load RSW file")
-		return
-	}
-
-	rsw := app.previewRSW
-
-	// Basic info
-	imgui.Text(fmt.Sprintf("Version: %s", rsw.Version))
-	imgui.Separator()
-
-	// File references
-	if imgui.TreeNodeExStrV("File References", imgui.TreeNodeFlagsDefaultOpen) {
-		if rsw.GndFile != "" {
-			imgui.Text(fmt.Sprintf("GND: %s", rsw.GndFile))
-		}
-		if rsw.GatFile != "" {
-			imgui.Text(fmt.Sprintf("GAT: %s", rsw.GatFile))
-		}
-		if rsw.IniFile != "" {
-			imgui.Text(fmt.Sprintf("INI: %s", rsw.IniFile))
-		}
-		if rsw.SrcFile != "" {
-			imgui.Text(fmt.Sprintf("SRC: %s", rsw.SrcFile))
-		}
-		imgui.TreePop()
-	}
-
-	imgui.Separator()
-
-	// Water settings
-	if imgui.TreeNodeExStrV("Water Settings", imgui.TreeNodeFlagsDefaultOpen) {
-		imgui.Text(fmt.Sprintf("Level: %.2f", rsw.Water.Level))
-		imgui.Text(fmt.Sprintf("Type: %d", rsw.Water.Type))
-		imgui.Text(fmt.Sprintf("Wave Height: %.2f", rsw.Water.WaveHeight))
-		imgui.Text(fmt.Sprintf("Wave Speed: %.2f", rsw.Water.WaveSpeed))
-		imgui.Text(fmt.Sprintf("Wave Pitch: %.2f", rsw.Water.WavePitch))
-		imgui.Text(fmt.Sprintf("Anim Speed: %d", rsw.Water.AnimSpeed))
-		imgui.TreePop()
-	}
-
-	imgui.Separator()
-
-	// Light settings
-	if imgui.TreeNodeExStrV("Light Settings", imgui.TreeNodeFlagsDefaultOpen) {
-		imgui.Text(fmt.Sprintf("Longitude: %d", rsw.Light.Longitude))
-		imgui.Text(fmt.Sprintf("Latitude: %d", rsw.Light.Latitude))
-		imgui.Text(fmt.Sprintf("Diffuse: (%.2f, %.2f, %.2f)", rsw.Light.Diffuse[0], rsw.Light.Diffuse[1], rsw.Light.Diffuse[2]))
-		imgui.Text(fmt.Sprintf("Ambient: (%.2f, %.2f, %.2f)", rsw.Light.Ambient[0], rsw.Light.Ambient[1], rsw.Light.Ambient[2]))
-		imgui.Text(fmt.Sprintf("Shadow Opacity: %.2f", rsw.Light.Opacity))
-		imgui.TreePop()
-	}
-
-	imgui.Separator()
-
-	// Object statistics
-	counts := rsw.CountByType()
-	if imgui.TreeNodeExStrV("Objects", imgui.TreeNodeFlagsDefaultOpen) {
-		imgui.Text(fmt.Sprintf("Total: %d objects", len(rsw.Objects)))
-		imgui.Text(fmt.Sprintf("Models: %d", counts[formats.RSWObjectModel]))
-		imgui.Text(fmt.Sprintf("Lights: %d", counts[formats.RSWObjectLight]))
-		imgui.Text(fmt.Sprintf("Sounds: %d", counts[formats.RSWObjectSound]))
-		imgui.Text(fmt.Sprintf("Effects: %d", counts[formats.RSWObjectEffect]))
-		imgui.TreePop()
-	}
-
-	imgui.Separator()
-
-	// Model list (collapsible)
-	models := rsw.GetModels()
-	if len(models) > 0 {
-		if imgui.TreeNodeExStrV(fmt.Sprintf("Model List (%d)", len(models)), imgui.TreeNodeFlagsNone) {
-			// Use clipper for large lists
-			for i, model := range models {
-				if i > 100 {
-					imgui.Text(fmt.Sprintf("... and %d more", len(models)-100))
-					break
-				}
-				name := model.Name
-				if name == "" {
-					name = model.ModelName
-				}
-				// Convert EUC-KR to UTF-8 for Korean names
-				imgui.Text(fmt.Sprintf("%d: %s", i, euckrToUTF8(name)))
-			}
-			imgui.TreePop()
-		}
-	}
-
-	// Sound list (collapsible)
-	sounds := rsw.GetSounds()
-	if len(sounds) > 0 {
-		if imgui.TreeNodeExStrV(fmt.Sprintf("Sound List (%d)", len(sounds)), imgui.TreeNodeFlagsNone) {
-			for i, sound := range sounds {
-				if i > 50 {
-					imgui.Text(fmt.Sprintf("... and %d more", len(sounds)-50))
-					break
-				}
-				imgui.Text(fmt.Sprintf("%d: %s", i, euckrToUTF8(sound.File)))
-			}
-			imgui.TreePop()
-		}
-	}
-
-	// Light source list (collapsible)
-	lights := rsw.GetLights()
-	if len(lights) > 0 {
-		if imgui.TreeNodeExStrV(fmt.Sprintf("Light Sources (%d)", len(lights)), imgui.TreeNodeFlagsNone) {
-			for i, light := range lights {
-				if i > 50 {
-					imgui.Text(fmt.Sprintf("... and %d more", len(lights)-50))
-					break
-				}
-				imgui.Text(fmt.Sprintf("%d: %s (range: %.1f)", i, euckrToUTF8(light.Name), light.Range))
-			}
-			imgui.TreePop()
-		}
-	}
-
-	// Effect list (collapsible)
-	effects := rsw.GetEffects()
-	if len(effects) > 0 {
-		if imgui.TreeNodeExStrV(fmt.Sprintf("Effects (%d)", len(effects)), imgui.TreeNodeFlagsNone) {
-			for i, effect := range effects {
-				if i > 50 {
-					imgui.Text(fmt.Sprintf("... and %d more", len(effects)-50))
-					break
-				}
-				imgui.Text(fmt.Sprintf("%d: %s (ID: %d)", i, euckrToUTF8(effect.Name), effect.EffectID))
-			}
-			imgui.TreePop()
-		}
-	}
-
-	// Quadtree info
-	if len(rsw.Quadtree) > 0 {
-		imgui.Separator()
-		imgui.Text(fmt.Sprintf("Quadtree nodes: %d", len(rsw.Quadtree)))
-	}
 }
 
 // renderStatusBar renders the status bar at the bottom.
